Melhorias para trabalhar com Restful multi nivel

+++ src/Controller/RestController.php
@@ -3,26 +3,26 @@
 namespace Solutio\Controller;
 
 use Zend\Mvc\Controller\AbstractRestfulController,
-    Zend\View\Model\JsonModel,
-    Zend\Json;
+  Zend\View\Model\JsonModel,
+  Zend\Json;
 
 class RestController extends AbstractRestfulController
 {
   private	$service;
-  
+
   public function __construct(\Solutio\Doctrine\EntityService $service)
   {
     $this->service = $service;
   }
-  
+
   public function getService()
   {
     return $this->service;
   }
-  
-  public function getEntity()
+
+  public function getEntity($data = null)
   {
-    return $this->getService()->getEntity();
+    return $this->getService()->getEntity($data);
   }
 
   // Listar - GET
@@ -31,7 +31,7 @@ class RestController extends AbstractRestfulController
     $entity = $this->getEntity();
     $data		= $this->service->find(new $entity($this->getDataEntity()), $this->getParams(), $this->getFields());
     return new JsonModel([
-      'data'		=> $data, 
+      'data'		=> $data,
       'success'	=> true
     ]);
   }
@@ -39,6 +39,13 @@ class RestController extends AbstractRestfulController
   // Retornar o registro especifico - GET
   public function get($id)
   {
+    $values = $this->getDataEntity();
+    if(count($values) > 0){
+      $ids  = $this->getEntity()::NameOfPrimaryKeys();
+      foreach($ids as $key)
+        if(empty($values[$key])) $values[$key] = $id;
+      $id = $values;
+    }
     $data = $this->service->getById($id);
     if($data){
       $return = [
@@ -47,8 +54,8 @@ class RestController extends AbstractRestfulController
       ];
     }else{
       $return = [
-        'message'	=> 'Objeto não encontrado.',
-        'success'	=> false
+        'success'	=> false,
+        'message'	=> 'The entity don\'t extists.'
       ];
     }
     return new JsonModel($return);
@@ -57,12 +64,16 @@ class RestController extends AbstractRestfulController
   // Insere registro - POST
   public function create($data)
   {
-    if($content = $this->getRequest()->getContent()){
-      $data = Json\Decoder::decode($content, Json\Json::TYPE_ARRAY);
+    $values = new \Solutio\Utils\Data\ArrayObject((array) $this->getDataEntity());
+    if(is_array($data) && count($data) > 0){
+      $data = new \Solutio\Utils\Data\ArrayObject((array) $data);
+    }else{
+      $data = new \Solutio\Utils\Data\ArrayObject(Json\Decoder::decode($this->getRequest()->getContent(), Json\Json::TYPE_ARRAY));
     }
-    if($data){
-      $entity = $this->getEntity();
-      $obj = $this->service->insert(new $entity($data));
+    $data   = $data->concat($values);
+    if($data->length() > 0){
+      $entity = $this->getEntity((array) $data);
+      $obj = $this->service->insert(new $entity((array) $data));
       if($obj)				{
         return new JsonModel([
           'data'		=> $obj,
@@ -78,16 +89,25 @@ class RestController extends AbstractRestfulController
   // alteracao - PUT
   public function update($id, $data)
   {
-    if($content = $this->getRequest()->getContent()){
-      $data = Json\Decoder::decode($content, Json\Json::TYPE_ARRAY);
+    $values = new \Solutio\Utils\Data\ArrayObject((array) $this->getDataEntity());
+    if(is_array($data) && count($data) > 0){
+      $data = new \Solutio\Utils\Data\ArrayObject((array) $data);
+    }else{
+      $data = new \Solutio\Utils\Data\ArrayObject(Json\Decoder::decode($data, Json\Json::TYPE_ARRAY));
     }
-    if($data){
-      if(!empty($id)){
-        $data = new \Solutio\Utils\Data\ArrayObject($data);
-        $data = (array) $data->concat(is_array($id) ? $id : ['id' => is_numeric($id) ? $id+0 : $id]);
-      }
+    $ids    = $this->getEntity()::NameOfPrimaryKeys();
+    if(count($ids) > 1){
+      $idsNotNull = 0;
+      foreach($ids as $key)
+        if(empty($values[$key])) $values[$key] = $id;
+        else $idsNotNull++;
+      if($idsNotNull < count($ids)-1)
+        $values = ['id' => $id];
+    }
+    $data   = $data->concat($values);
+    if($data->length() > 0){
       $entity = $this->getEntity();
-      $entity = new $entity($data);
+      $entity = new $entity((array) $data);
       $obj = $this->service->update($entity);
       if($obj){
         return new JsonModel([
@@ -104,20 +124,32 @@ class RestController extends AbstractRestfulController
   // delete - DELETE
   public function delete($id)
   {
-    $data = new \Solutio\Utils\Data\ArrayObject();
-    $data = (array) $data->concat(is_array($id) ? $id : ['id' => $id]);
-    $entity = $this->getEntity();
-    $entity = new $entity($data);
-    $res = $this->service->delete($entity);
-    if($res){
-      return new JsonModel(['success' => true]);
-    }else
-      return new JsonModel(['success' => false]);
+    $values = new \Solutio\Utils\Data\ArrayObject((array) $this->getDataEntity());
+    $data   = new \Solutio\Utils\Data\ArrayObject;
+    if($content = $this->getRequest()->getContent()){
+      $data = $data->concat(Json\Decoder::decode($content, Json\Json::TYPE_ARRAY));
+    }
+    if(count($values) > 0){
+      $ids  = $this->getEntity()::NameOfPrimaryKeys();
+      foreach($ids as $key)
+        if(empty($values[$key])) $values[$key] = $id;
+    }
+    $data   = $data->concat($values);
+    if($data->length() > 0){
+      $entity = $this->getEntity();
+      $entity = new $entity((array) $data);
+      $res = $this->service->delete($entity);
+      if($res){
+        return new JsonModel(['success' => true]);
+      }else
+        return new JsonModel(['success' => false]);
+    }
+    return new JsonModel(['success' => false]);
   }
-  
-  protected function getDataEntity()
+
+  protected function getDataEntity(array $data = [])
   {
-    $data = $this->getRequest()->getQuery()->toArray();
+    $data = array_merge($this->getRequest()->getQuery()->toArray(), $data);
     foreach($data as $k => $v){
       if(!empty($v)){
         $data[$k] = json_decode($v, true);
@@ -127,7 +159,7 @@ class RestController extends AbstractRestfulController
     }
     return $data;
   }
-  
+
   protected function getParams()
   {
     $get		= $this->getRequest()->getQuery();
@@ -137,13 +169,13 @@ class RestController extends AbstractRestfulController
       'order'		=> json_decode($get->get('order'), true)
     ];
   }
-  
+
   protected function getFields()
   {
     $get		= $this->getRequest()->getQuery();
     $fields		= [];
     if($get->get('fields'))
-      $fields		= explode(',', $get->get('fields'));	
+      $fields		= explode(',', $get->get('fields'));
     return $fields;
   }
 }
